implemented sqlite as a storage engine created exceptions holder started logger class
git-svn-id: https://zenses.googlecode.com/svn/trunk@39 c5ca1edb-0f35-0410-b988-7bac8ecc5e99

+++ src/zensesLib/Managers/EntryManager.cs
@@ -106,7 +106,7 @@ namespace Zenses.Lib.Managers
 
 				if (properties[PropertyNames.OBJECT_DURATION] != null)
 				{
-					entry.Length = (long)(properties[PropertyNames.OBJECT_DURATION]);
+					entry.Length = (int)((long)properties[PropertyNames.OBJECT_DURATION]);
 				}
 				else
 				{
